fix: added version to spec config to check for changes between spec file versions
Some versions of spec files are updated and "fix" previous issues.  For example, inputs.conf.spec does not define the interval parameter for WinPrintMon stanzas prior to version 9.  By implementing a version in the spec config object, we can check for these one-offs and address them in code.  This fix addresses issue #53